+++ src/core/bullet-utils.ts
@@ -1,16 +1,14 @@
-import { Bullet } from "src"
+import { Bullet } from "./bullet-types";
 
 export function updateBulletType(original: string, newType: Bullet): string {
   if (!isBulletText(original)) {
-    // raise error
-    throw new Error('The provided text is not a valid bullet point.')
+    throw new Error("The provided text is not a valid bullet point.");
   }
 
-  return original.replace(/- \[.\]/, `- [${newType.character}]`)
+  return original.replace(/- \[.\]/, `- [${newType.character}]`);
 }
 
 export function isBulletText(text: string): boolean {
-  // trimmed text starts with - [
-  const trimmed = text.trimStart()
-  return trimmed.startsWith('- [')
+  const trimmed = text.trimStart();
+  return trimmed.startsWith("- [");
 }